refactor: inline S3 form cleanup
Keep the credential-clearing fix in its single caller and remove testing-only files from the PR.
